修改权限验证，无法鉴权进行跳转，修改对应js

+++ src/com/common/ShiroDbRealm.java
@@ -90,7 +90,7 @@ public class ShiroDbRealm extends AuthorizingRealm {
     	if (uai==null) {
 			return null;
 		}
-    	Integer userId = (Integer) principals.fromRealm(getName()).iterator().next();
+    	Long userId = (Long) principals.fromRealm(getName()).iterator().next();
     	return uai.getAuthorizationById(userId.toString());
     }
     /**
